+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-common.spec.tsx
@@ -113,7 +113,7 @@ describe('Patient dashboard for anyone', () => {
 
     const medicalFilesCard = screen.getByTestId('medical-files-card')
     expect(medicalFilesCard).toBeVisible()
-    expect(medicalFilesCard).toHaveTextContent('Medical filesPrescriptionsPrescription_2022-01-02Medical reportsMedical report 2022-01-02Medical report 2022-01-02_1New')
+    expect(medicalFilesCard).toHaveTextContent('Medical filesMedical report-1 2022-01-10Created by Vishnou LapaixMySecondTeamMedical report-2 2022-01-02Created by Vishnou LapaixMySecondTeamNew')
 
     const alarmCard = screen.getByTestId('alarm-card')
     expect(alarmCard).toBeVisible()