+++ CHANGELOG.md
@@ -8,6 +8,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2140 As a HCP, I can only edit my medical records
 - YLP-2160 Patient list new design
 - YLP-2175 Implement new design system
+- YLP-2231 Implement new medical reports design for patients
 
 ### Engineering use
 - YLP-2044 Remove redux settings store